\docType{package}
\name{cocron-package}
\alias{cocron-package}
\title{The cocron Package}
\description{
  Statistical comparisons of two or more alpha
  coefficients.
}
\details{
  \tabular{ll}{ Package: \tab cocron\cr Type: \tab
  Package\cr Version: \tab 1.0-0\cr Date: \tab
  2013-09-05\cr Depends: \tab methods\cr Encoding: \tab
  UTF-8\cr License: \tab GPL (>= 3)\cr LazyLoad: \tab
  yes\cr URL: \tab
  http://r.birkdiedenhofen.de/pckg/cocron/\cr }

  This package provides functions to statistically compare
  two or more alpha coefficients based on either dependent
  or independent groups of individuals.
}
\author{
  Birk Diedenhofen \email{mail@birkdiedenhofen.de}
}
\keyword{package}

