\encoding{UTF-8}
\name{knowledge}
\Rdversion{1.1}
\alias{knowledge}
\docType{data}
\title{
Sample dataset: knowledge
}
\description{
Data of 312 testees who completed two tests on general knowledge consisting of 30 questions each.
}
\usage{data(knowledge)}
\format{
  A list that contains a matrix for each of the two tests holding 312 observations (rows) on the 30 questions (columns) presented. The ones and zeros stand for correct and incorrect responses, respectively.
}
\examples{
data(knowledge)
}
\keyword{datasets}
