% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocron-package.R
\docType{package}
\name{cocron-package}
\alias{cocron-package}
\title{The cocron Package}
\description{
Statistical Comparisons of Two or more Alpha Coefficients.
}
\details{
\tabular{ll}{
Package: \tab cocron\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2016-03-11\cr
Depends: \tab methods\cr
Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://comparingcronbachalphas.org\cr
}

Statistical tests for the comparison between two or more alpha coefficients based on either dependent or independent groups of individuals. A web interface is available at http://comparingcronbachalphas.org. A plugin for the R GUI and IDE RKWard is included. Please install RKWard from https://rkward.kde.org to use this feature. The respective R package 'rkward' cannot be installed directly from a repository,
      as it is a part of RKWard.
}
\author{
Birk Diedenhofen
}
\keyword{package}

